refactor: divide folder for interest
- divide with Model, View
