+++ src/Model/Level/LevelData.java
@@ -1,4 +1,4 @@
-/*
+package Model.Level;/*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
