+++ src/Model/Plant/Plant.java
@@ -1,3 +1,7 @@
+package Model.Plant;
+
+import View.Game.GamePanel;
+
 /**
  * Created by Armin on 6/25/2016.
  */
