+++ src/Model/Plant/PlantButton.java
@@ -1,3 +1,5 @@
+package Model.Plant;
+
 /**
  * Created by Armin on 6/25/2016.
  */
