+++ src/View/Collider.java
@@ -1,3 +1,7 @@
+package View;
+
+import Model.Plant.Plant;
+
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
