+++ src/View/Element/PlantCard.java
@@ -1,3 +1,5 @@
+package View.Element;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
