+++ src/View/Element/Sun.java
@@ -1,3 +1,7 @@
+package View.Element;
+
+import View.Game.GamePanel;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseEvent;
@@ -25,7 +29,7 @@ public class Sun extends JPanel implements MouseListener {
         myX = startX;
         myY = startY;
         setLocation(myX, myY);
-        sunImage = new ImageIcon(this.getClass().getResource("images/sun.png")).getImage();
+        sunImage = new ImageIcon(this.getClass().getResource("../../images/sun.png")).getImage();
         addMouseListener(this);
     }
